requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
google-cloud-storage==2.18.2
google-auth==2.34.0
